refactor: insert BIS notifications one row at a time for exact insert ids
