+++ plugins/woocommerce/src/Internal/StockNotifications/Migration/Writers/DbWriter.php
@@ -15,35 +15,39 @@ defined( 'ABSPATH' ) || exit;
 /**
  * Live writer for the Back In Stock Notifications migration.
  *
- * Notification inserts are delegated to `BulkNotificationWriter`. Meta written onto an
- * existing notification — natural-key adoption markers and legacy unsubscribe tokens — is
- * always inserted, never updated, and goes through direct SQL rather than `add_meta_data()`,
- * which would bump `date_modified_gmt` on a row the merchant did not touch. `write_product_meta()`
- * is the one exception, going through the product CRUD layer per the plan.
+ * Notifications are never routed through `Notification` + `save()`. Each row is inserted on
+ * its own so the id it was given comes straight back from `$wpdb->insert_id`, then its meta
+ * goes in with one multi-row statement against that id. The whole call runs inside a single
+ * transaction: a failure between a row and its meta would leave a Core row with no
+ * `_wc_bis_legacy_id` marker, invisible to the candidate predicate and re-inserted on the
+ * next run.
+ *
+ * Meta written onto an existing notification — natural-key adoption markers and legacy
+ * unsubscribe tokens — is always inserted, never updated, and goes through direct SQL rather
+ * than `add_meta_data()`, which would bump `date_modified_gmt` on a row the merchant did not
+ * touch. `write_product_meta()` is the one exception, going through the product CRUD layer
+ * per the plan.
  */
 class DbWriter implements WriterInterface {
 
 	/**
-	 * Bulk insert engine for `wc_stock_notifications` and its meta.
+	 * Columns accepted on `wc_stock_notifications`, mapped to their `$wpdb->prepare()` format.
 	 *
-	 * @var BulkNotificationWriter
+	 * @var array<string,string>
 	 */
-	private BulkNotificationWriter $bulk_writer;
-
-	/**
-	 * Initialize the class instance.
-	 *
-	 * Resolve this writer from the container rather than constructing it directly: built with
-	 * `new`, `$bulk_writer` is never set and the first insert fails on an uninitialized
-	 * typed property.
-	 *
-	 * @internal
-	 * @param BulkNotificationWriter $bulk_writer Bulk insert engine for the notifications table.
-	 * @return void
-	 */
-	final public function init( BulkNotificationWriter $bulk_writer ): void {
-		$this->bulk_writer = $bulk_writer;
-	}
+	private const COLUMN_FORMATS = array(
+		'product_id'            => '%d',
+		'user_id'               => '%d',
+		'user_email'            => '%s',
+		'status'                => '%s',
+		'date_created_gmt'      => '%s',
+		'date_modified_gmt'     => '%s',
+		'date_confirmed_gmt'    => '%s',
+		'date_last_attempt_gmt' => '%s',
+		'date_notified_gmt'     => '%s',
+		'date_cancelled_gmt'    => '%s',
+		'cancellation_source'   => '%s',
+	);
 
 	/**
 	 * This writer performs real writes.
@@ -55,13 +59,80 @@ class DbWriter implements WriterInterface {
 	}
 
 	/**
-	 * Insert notifications together with their meta, in chunks, one transaction per chunk.
+	 * Insert notifications together with their meta, inside a single transaction.
 	 *
 	 * @param array $rows List of rows, each `array{ columns: array<string,mixed>, meta: array<int,array{0:string,1:mixed}> }`.
+	 * @throws \Exception If any insert fails; the whole call is rolled back first.
 	 * @return int Number of notifications written.
 	 */
 	public function insert_notifications( array $rows ): int {
-		return $this->bulk_writer->insert_notifications( $rows );
+		global $wpdb;
+
+		if ( empty( $rows ) ) {
+			return 0;
+		}
+
+		$written = 0;
+
+		$wpdb->query( 'START TRANSACTION' );
+
+		try {
+			foreach ( $rows as $row ) {
+				$notification_id = $this->insert_notification_row( $row['columns'] );
+				$meta            = $row['meta'] ?? array();
+
+				if ( ! empty( $meta ) ) {
+					$this->write_notification_meta( $notification_id, $meta );
+				}
+
+				++$written;
+			}
+		} catch ( \Exception $e ) {
+			$wpdb->query( 'ROLLBACK' );
+			throw $e;
+		}
+
+		$wpdb->query( 'COMMIT' );
+
+		return $written;
+	}
+
+	/**
+	 * Insert one row into the notifications table and return the id it was given.
+	 *
+	 * A column the caller left out is written as SQL NULL rather than skipped: `$wpdb->insert()`
+	 * emits a NULL literal for a null value, where a missing column would fall back to whatever
+	 * the schema defaults to. Core reads the date columns with `IS NULL`, so a zero date would
+	 * read as "already attempted".
+	 *
+	 * @param array $columns Column values for this row.
+	 * @throws \RuntimeException If the insert fails or reports no id.
+	 * @return int Notification id.
+	 */
+	private function insert_notification_row( array $columns ): int {
+		global $wpdb;
+
+		$data    = array();
+		$formats = array();
+
+		foreach ( self::COLUMN_FORMATS as $column => $format ) {
+			$data[ $column ] = $columns[ $column ] ?? null;
+			$formats[]       = $format;
+		}
+
+		$result = $wpdb->insert( Tables::core_notifications(), $data, $formats );
+
+		if ( false === $result ) {
+			throw new \RuntimeException( 'Failed to insert stock notification row: ' . $wpdb->last_error ); // phpcs:ignore WordPress.Security.EscapeOutput.ExceptionNotEscaped
+		}
+
+		$notification_id = (int) $wpdb->insert_id;
+
+		if ( $notification_id <= 0 ) {
+			throw new \RuntimeException( 'Inserted a stock notification row but no id was reported.' );
+		}
+
+		return $notification_id;
 	}
 
 	/**
@@ -75,11 +146,31 @@ class DbWriter implements WriterInterface {
 	 * @return int Number of meta rows written.
 	 */
 	public function insert_notification_meta( int $notification_id, array $meta ): int {
-		global $wpdb;
-
 		if ( empty( $meta ) ) {
 			return 0;
 		}
+
+		try {
+			$written = $this->write_notification_meta( $notification_id, $meta );
+		} catch ( \RuntimeException $e ) {
+			return 0;
+		}
+
+		$this->invalidate_meta_cache( $notification_id );
+
+		return $written;
+	}
+
+	/**
+	 * Write one notification's meta rows with a single multi-row statement.
+	 *
+	 * @param int   $notification_id Target notification id.
+	 * @param array $meta            List of `array{0:string,1:mixed}` key/value pairs.
+	 * @throws \RuntimeException If the insert fails.
+	 * @return int Number of meta rows written.
+	 */
+	private function write_notification_meta( int $notification_id, array $meta ): int {
+		global $wpdb;
 
 		$meta_table = Tables::core_meta();
 		$values     = array();
@@ -87,8 +178,8 @@ class DbWriter implements WriterInterface {
 		foreach ( $meta as $pair ) {
 			$values[] = $notification_id;
 			$values[] = $pair[0];
-			// This is the sole owner of maybe_serialize() for meta values, matching
-			// BulkNotificationWriter; callers must hand over unserialized values.
+			// This writer is the sole owner of maybe_serialize() for meta values; callers
+			// (e.g. MetaMapper) must hand over unserialized values, or this double-serializes.
 			$values[] = maybe_serialize( $pair[1] );
 		}
 
@@ -104,10 +195,8 @@ class DbWriter implements WriterInterface {
 		// phpcs:enable WordPress.DB.PreparedSQL.InterpolatedNotPrepared, WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.PreparedSQLPlaceholders.UnfinishedPrepare
 
 		if ( false === $result ) {
-			return 0;
+			throw new \RuntimeException( 'Failed to insert stock notification meta rows: ' . $wpdb->last_error ); // phpcs:ignore WordPress.Security.EscapeOutput.ExceptionNotEscaped
 		}
-
-		$this->invalidate_meta_cache( $notification_id );
 
 		return (int) $result;
 	}
